requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.37.3
botocore==1.37.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
mysql-connector==2.2.9
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openai==1.65.1
openpyxl==3.1.5
oracledb==2.5.0
packaging==24.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
redshift-connector==2.1.5
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.11.3
scramp==1.4.5
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sshtunnel==0.4.0
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
